refactor: move /upload-docs to /docs/upload
Ref. metriport/metriport-internal#704
