+++ packages/api/src/routes/medical/internal-docs.ts
@@ -1,7 +1,9 @@
 import { Request, Response } from "express";
 import Router from "express-promise-router";
 import httpStatus from "http-status";
+import multer from "multer";
 import { z } from "zod";
+import { createAndUploadDocReference } from "../../command/medical/admin/upload-doc";
 import { checkDocumentQueries } from "../../command/medical/document/check-doc-queries";
 import {
   isDocumentQueryProgressEqual,
@@ -12,15 +14,23 @@ import { getPatientOrFail } from "../../command/medical/patient/get-patient";
 import { MAPIWebhookStatus, processPatientDocumentRequest } from "../../command/webhook/medical";
 import { convertResult } from "../../domain/medical/document-query";
 import BadRequestError from "../../errors/bad-request";
+import { makeS3Client } from "../../external/aws/s3";
+import { Config } from "../../shared/config";
+import { createS3FileName } from "../../shared/external";
 import { capture } from "../../shared/notifications";
 import { stringToBoolean } from "../../shared/types";
 import { Util } from "../../shared/util";
+import { uuidv7 } from "../../shared/uuid-v7";
 import { documentQueryProgressSchema } from "../schemas/internal";
 import { stringListSchema } from "../schemas/shared";
 import { getUUIDFrom } from "../schemas/uuid";
 import { asyncHandler, getFrom } from "../util";
+import { getFromQueryOrFail } from "./../util";
 
 const router = Router();
+const upload = multer();
+const s3client = makeS3Client();
+const bucketName = Config.getMedicalDocumentsBucketName();
 
 /** ---------------------------------------------------------------------------
  * POST /internal/docs/reprocess
@@ -176,6 +186,61 @@ router.post(
     const patientIds = patientIdsRaw?.split(",") ?? [];
     checkDocumentQueries(patientIds);
     return res.sendStatus(httpStatus.ACCEPTED);
+  })
+);
+
+/** ---------------------------------------------------------------------------
+ * POST /internal/docs/upload
+ *
+ * Upload doc for a patient.
+ * 
+ * Originally on packages/api/src/routes/internal.ts
+ *
+ * @param req.query.cxId - The customer/account's ID.
+ * @param req.query.patientId - The patient ID.
+ * @param req.file - The file to be stored.
+ * @param req.body.metadata - The metadata for the file.
+
+ * @return 200 Indicating the file was successfully uploaded.
+ */
+router.post(
+  "/upload",
+  upload.single("file"),
+  asyncHandler(async (req: Request, res: Response) => {
+    const cxId = getUUIDFrom("query", req, "cxId").orFail();
+    const patientId = getFromQueryOrFail("patientId", req);
+    const file = req.file;
+
+    if (!file) {
+      throw new BadRequestError("File must be provided");
+    }
+
+    const docRefId = uuidv7();
+    const fileName = createS3FileName(cxId, patientId, docRefId);
+
+    await s3client
+      .upload({
+        Bucket: bucketName,
+        Key: fileName,
+        Body: file.buffer,
+        ContentType: file.mimetype,
+      })
+      .promise();
+
+    const metadata = JSON.parse(req.body.metadata);
+
+    const docRef = await createAndUploadDocReference({
+      cxId,
+      patientId,
+      docId: docRefId,
+      file: {
+        ...file,
+        originalname: fileName,
+      },
+      metadata,
+    });
+
+    return res.status(httpStatus.OK).json(docRef);
   })
 );
